Made cmd variable a byte for binary operations
This variable should have never been an int since it is defined in
the protocol as a byte. The reason for the change is that when we
have opcode that is greater have 127 the value of a byte and an int
are different (Ex. int = 128 byte = -127). This was causing an
assertion to fail, in this case for the getl command (0x94).

Reviewed-on: http://review.couchbase.org/9476
Tested-by: Michael Wiederhold <mike@couchbase.com>

+++ src/main/java/net/spy/memcached/protocol/binary/DeleteOperationImpl.java
@@ -6,7 +6,7 @@ import net.spy.memcached.ops.OperationCallback;
 class DeleteOperationImpl extends SingleKeyOperationImpl implements
 		DeleteOperation {
 
-	private static final int CMD=0x04;
+	private static final byte CMD=0x04;
 
 	private final long cas;
