+++ src/main/java/net/spy/memcached/protocol/binary/GetAndTouchOperationImpl.java
@@ -5,7 +5,7 @@ import net.spy.memcached.ops.GetAndTouchOperation;
 public class GetAndTouchOperationImpl extends SingleKeyOperationImpl
 		implements GetAndTouchOperation {
 
-	static final int GAT_CMD=0x1d;
+	static final byte GAT_CMD=0x1d;
 
 	/**
 	 * Length of the extra header stuff for a GET response.
