+++ src/main/java/net/spy/memcached/protocol/binary/MutatorOperationImpl.java
@@ -8,8 +8,8 @@ import net.spy.memcached.ops.OperationStatus;
 class MutatorOperationImpl extends SingleKeyOperationImpl implements
 		MutatorOperation {
 
-	private static final int CMD_INCR=0x05;
-	private static final int CMD_DECR=0x06;
+	private static final byte CMD_INCR=0x05;
+	private static final byte CMD_DECR=0x06;
 
 	private final Mutator mutator;
 	private final long by;
